Rename parser to craig-like parser
