+++ src/y0/parser/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+"""Implementation of the y0 probability expression grammar with :mod:`pyparsing`."""
+
+from .craig.grammar import parse_craig  # noqa:F401
