+++ src/y0/parser/craig/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""A parser for Craig-like probability expressions based on :mod:`pyparsing`."""
